fix(checkout): CHECKOUT-10026 Move service selection logic to a separate hook
