+++ packages/core/src/app/address/googleAutocomplete/utils.ts
@@ -1,0 +1,12 @@
+import { type AutocompleteItem } from '@bigcommerce/checkout/ui';
+
+export const toAutocompleteItems = (
+    results?: google.maps.places.AutocompletePrediction[],
+): AutocompleteItem[] => {
+    return (results || []).map((result) => ({
+        label: result.description,
+        value: result.structured_formatting.main_text,
+        highlightedSlices: result.matched_substrings,
+        id: result.place_id,
+    }));
+};
